Dynamically detects relevant Spark methods and analyzes them
INVOKEVIRTUAL instructions are intercepted to detect calls to
relevant Spark methods (At this point: filter, map and reduce).

The methods are inspected and a String is built that represents
the method to be analyzed. In the case of anonymous classes
a representation of this one is taken from the previous element
in the stack and in the case of lambdas, a call to the private
synthetic method of the form lambda${x} where {x} represent a
0 based sequence is created.