Refactors the code of the coordinator
The aim of this refactor is to be able to adapt
to the right strategy depending on the spark method
invoked. For example, in the case of filters, the
initial symbolic value most be kept and forwarded to
the following operations. In the case of map, the
transformed symbolic value is the one that should be
kept and forwarded.