Adds support to lambdas and also tries to set the correct strategy
When backtracking, the right strategy is chosen again to avoid dropping
some paths by mistake. Also, validations are only added when the last
method is reached. It still fails to add the value when a path is cut
out like in the case of a negative branch in the filter.