Solves the bug of not detecting correctly chained mpas and filters
Initially some chained methods were not behaving correctly because
when backtracking the proper method strategy was not correctly
set. For example, if a map and a filter were executed in that order
then whe backtracking to the other choices in the map the method
strategy used would be the one for filters (last method invoked)

For this reason, some paths were dropped, and some invalid values
were added to the result.