Includes support to the FlatMap action
This caused a lot of refactoring. For starters, I returned to the
stateless strategies approach because keeping states in the previously
used strategies would generate problems when backtracking with the new
choice generator for the multiple flatMap values.

Also now the strategies have an inputExpression and a list of outputExpressions.

The first element in the list of outputExpressions is passed as the input
to the next strategy if any.