Updates the mocked Spark library to support Word Count
The word count example introduces some operations that were not mocked until now.

For example, mapToPair to create JavaPairRDD that is an RDD of Tuple2 elements. Also
the necessary transformations in JavaPairRDD were created. For example reduceByKey.

It is important to note that this operations are not analyzed yet because they
are executed on objects (Tuple2) that are not yet processed symbolically in this
implementation.